refactor: add get address APDU command handler and UI flow
This commit introduces the get address functionality with several key improvements:
- Implement handleGetAddress function in get_address.c
- Update dispatcher to support INS_GET_ADDRESS command
- Add address validation and UI confirmation flow
- Modify send_response helper to return int for consistent error handling
- Update UI display logic to use a generic validation callback
- Add validate_address function for address export and rejection

+++ src/helper/send_response.h
@@ -4,6 +4,6 @@
 #include "io.h"
 
 int helper_send_response_public_key();
-uint8_t helper_send_response_address();
+int helper_send_response_address();
 uint8_t helper_send_response_sign_transaction();
 uint8_t helper_send_response_sign();
